main/requirements.txt
requests
flask
matplotlib
pandas
pillow